+++ src/Axent.Extensions.Caching/AxentCacheBuilderExtensions.cs
@@ -1,0 +1,16 @@
+using Microsoft.Extensions.DependencyInjection;
+
+namespace Axent.Extensions.Caching;
+
+public static class AxentCacheBuilderExtensions
+{
+    extension(IAxentCacheBuilder builder)
+    {
+        public IAxentCacheBuilder UseInMemory()
+        {
+            builder.Services.AddMemoryCache();
+            builder.Services.AddSingleton<ICache, InMemoryCache>();
+            return builder;
+        }
+    }
+}