+++ src/Axent.Extensions.Caching/IAxentCacheBuilder.cs
@@ -1,0 +1,8 @@
+using Microsoft.Extensions.DependencyInjection;
+
+namespace Axent.Extensions.Caching;
+
+public interface IAxentCacheBuilder
+{
+    IServiceCollection Services { get; }
+}